completed separating of operations protocols and extensions, fixed spacing and formatting error in the shared new file header template macro
