+++ Source/SwiftCloudant/Operations/Operation.swift
@@ -1,5 +1,6 @@
 //
 //  Operation.swift
+//
 //  SwiftCloudant
 //
 //  Created by Rhys Short on 03/03/2016.
