+++ SwiftCloudant.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,9 @@
 <Workspace
    version = "1.0">
    <FileRef
+      location = "group:Source/SwiftCloudant/TestFile.swift">
+   </FileRef>
+   <FileRef
       location = "group:Tests/SwiftCloudantTests/SwiftCloudantTestExample.swift">
    </FileRef>
    <FileRef
